+++ src/scriptlets/remove-cookie.js
@@ -95,6 +95,10 @@ removeCookie.names = [
     'cookie-remover.js',
     'ubo-cookie-remover.js',
     'ubo-cookie-remover',
+    'remove-cookie.js',
+    'ubo-remove-cookie.js',
+    'ubo-remove-cookie',
+    'abp-cookie-remover',
 ];
 
 removeCookie.injections = [toRegExp, hit];